ajustes nos diretorios
